Task "Lord of the Rings RPG" solution. Changed packages, divided into sprints.
