--- src/main/java/school/faang/bjs2_32707/Student.java
@@ -1,16 +1,0 @@
-package school.faang.bjs2_32707;
-
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.ToString;
-
-@Getter
-@AllArgsConstructor
-@ToString
-@EqualsAndHashCode
-public class Student {
-    private String name;
-    private String faculty;
-    private int year;
-}
