--- src/main/java/school/faang/bjs2_32814/WebPage.java
@@ -1,11 +1,0 @@
-package school.faang.bjs2_32814;
-
-
-import lombok.Data;
-
-@Data
-public class WebPage {
-    private final String url;
-    private final String title;
-    private final String content;
-}
