--- src/main/java/school/faang/bjs2_33038/Book.java
@@ -1,16 +1,0 @@
-package school.faang.bjs2_33038;
-
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.ToString;
-
-@EqualsAndHashCode
-@AllArgsConstructor
-@ToString
-@Getter
-public class Book {
-    private final String title;
-    private final String author;
-    private final int year;
-}
