--- src/main/java/school/faang/bjs2_33768/SpellAction.java
@@ -1,6 +1,0 @@
-package school.faang.bjs2_33768;
-
-@FunctionalInterface
-public interface SpellAction {
-    String castSpell(String spellName);
-}
