--- src/main/java/school/faang/bjs2_33768/SpellCaster.java
@@ -1,8 +1,0 @@
-package school.faang.bjs2_33768;
-
-public class SpellCaster {
-    public void cast(String spellName, SpellAction spellAction) {
-        String result = spellAction.castSpell(spellName);
-        System.out.println("Результат заклинание: " + result);
-    }
-}
