--- src/main/java/school/faang/bjs2_34765/Image.java
@@ -1,13 +1,0 @@
-package school.faang.bjs2_34765;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.ToString;
-
-@Getter
-@AllArgsConstructor
-@ToString
-public class Image {
-    private String name;
-    private String description;
-}
