--- src/main/java/school/faang/bsj2_33230/StreamEvent.java
@@ -1,16 +1,0 @@
-package school.faang.bsj2_33230;
-
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.ToString;
-
-@AllArgsConstructor
-@ToString
-@EqualsAndHashCode
-@Getter
-public class StreamEvent {
-    private int id;
-    private String eventType;
-    private String data;
-}
