+++ src/main/java/school/faang/sprint_1/bjs2_32707/Student.java
@@ -1,0 +1,16 @@
+package school.faang.sprint_1.bjs2_32707;
+
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@AllArgsConstructor
+@ToString
+@EqualsAndHashCode
+public class Student {
+    private String name;
+    private String faculty;
+    private int year;
+}
