+++ src/main/java/school/faang/sprint_1/bjs2_32814/WebPage.java
@@ -1,0 +1,11 @@
+package school.faang.sprint_1.bjs2_32814;
+
+
+import lombok.Data;
+
+@Data
+public class WebPage {
+    private final String url;
+    private final String title;
+    private final String content;
+}
