+++ src/main/java/school/faang/sprint_1/bsj2_33230/StreamEvent.java
@@ -1,0 +1,16 @@
+package school.faang.sprint_1.bsj2_33230;
+
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.ToString;
+
+@AllArgsConstructor
+@ToString
+@EqualsAndHashCode
+@Getter
+public class StreamEvent {
+    private int id;
+    private String eventType;
+    private String data;
+}
