+++ src/main/java/school/faang/sprint_2/bjs2_33712/Character.java
@@ -1,0 +1,17 @@
+package school.faang.sprint_2.bjs2_33712;
+
+import lombok.Getter;
+
+import java.util.HashSet;
+import java.util.Set;
+
+@Getter
+public class Character {
+    private final String name;
+    private final Set<Item> inventory;
+
+    Character(String name) {
+        this.name = name;
+        this.inventory = new HashSet<>();
+    }
+}
