+++ src/main/java/school/faang/sprint_2/bjs2_33712/InventoryManager.java
@@ -1,0 +1,23 @@
+package school.faang.sprint_2.bjs2_33712;
+
+import java.util.function.Consumer;
+import java.util.function.Predicate;
+
+public class InventoryManager {
+    public void addItem(Character character, Item item, Consumer<Item> consumer) {
+        character.getInventory().add(item);
+        consumer.accept(item);
+    }
+
+    public void removeItem(Character character, Predicate<Item> predicate) {
+        character.getInventory().removeIf(predicate);
+    }
+
+    public void updateItemValue(Character character, Predicate<Item> predicate, Consumer<Item> updater) {
+        character.getInventory().forEach(item -> {
+            if (predicate.test(item)) {
+                updater.accept(item);
+            }
+        });
+    }
+}
