+++ src/main/java/school/faang/sprint_2/bjs2_33712/Item.java
@@ -1,0 +1,15 @@
+package school.faang.sprint_2.bjs2_33712;
+
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.Setter;
+
+@AllArgsConstructor
+@Getter
+@Setter
+@EqualsAndHashCode
+public class Item {
+    private String name;
+    private int value;
+}
