+++ src/main/java/school/faang/sprint_2/bjs2_33768/SpellAction.java
@@ -1,0 +1,6 @@
+package school.faang.sprint_2.bjs2_33768;
+
+@FunctionalInterface
+public interface SpellAction {
+    String castSpell(String spellName);
+}
