+++ src/main/java/school/faang/sprint_2/bjs2_33768/SpellCaster.java
@@ -1,0 +1,8 @@
+package school.faang.sprint_2.bjs2_33768;
+
+public class SpellCaster {
+    public void cast(String spellName, SpellAction spellAction) {
+        String result = spellAction.castSpell(spellName);
+        System.out.println("Результат заклинание: " + result);
+    }
+}
