+++ src/main/java/school/faang/sprint_2/bjs2_34765/FilterProcessor.java
@@ -1,0 +1,14 @@
+package school.faang.sprint_2.bjs2_34765;
+
+import java.util.function.Function;
+
+public class FilterProcessor {
+    public Image applyFilter(Image image, Function<Image, Image> filter) {
+        return filter.apply(image);
+    }
+
+    public Function<Image, Image> combineFilters(Function<Image, Image> firstFilter,
+                                                 Function<Image, Image> secondFilter) {
+        return image -> firstFilter.andThen(secondFilter).apply(image);
+    }
+}
