+++ src/main/java/school/faang/sprint_2/bjs2_34765/Image.java
@@ -1,0 +1,13 @@
+package school.faang.sprint_2.bjs2_34765;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.ToString;
+
+@Getter
+@AllArgsConstructor
+@ToString
+public class Image {
+    private String name;
+    private String description;
+}
